# openapi.yml
openapi: 3.0.3
info:
  title: iPets
  description: |-
    Especificacao dos contratos disponibilizados pelo servidor do aplicativo iPets, um
    aplicativo para contratacao de servicos para pets.
    Parte do projeto do grupo 21 na disciplina de Engenharia de Software, ministrada em
    2023.1 na UFABC.
  version: 0.1.0
servers:
  - url: http://localhost:8080/
...
paths:
...
  /services:
    put:
      tags:
        - Servicos
      summary: Cadastra os dados de um novo servico na plataforma
      operationId: createService
      requestBody:
        description: Dado do servico a ser cadastrado
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateServiceModel"
      responses:
        "200":
          description: Servico criado com sucesso
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServiceModel"
        "400":
          description: Servico requisitado possui dados invalidos
    get:
      tags:
        - Utilidades
      summary: Obtem todos os servicos existentes na plataforma
      operationId: getAllServices
      responses:
        "200":
          description: IDs obtidos com sucesso
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServiceListModel"
  /services/{objectId}:
    get:
      tags:
        - Servicos
      summary: Obtem os dados de um servico existente por ID na plataforma
      operationId: getServiceById
      parameters:
        - $ref: "#/components/parameters/ObjectId"
      responses:
        "200":
          description: Dados do servico obtidos com sucesso
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServiceModel"
        "404":
          description: Servico inexistente para o ID
    patch:
      tags:
        - Servicos
      summary: Atualiza os dados de um servico existente por ID na plataforma
      operationId: updateServiceById
      parameters:
        - $ref: "#/components/parameters/ObjectId"
      requestBody:
        description: Dado do servico a ser atualizado
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateServiceModel"
      responses:
        "200":
          description: Dados do servico atualizados com sucesso
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServiceModel"
        "404":
          description: Servico inexistente para o ID
    delete:
      tags:
        - Servicos
      summary: Remove os dados de um servico existente por ID na plataforma
      operationId: removeServiceById
      parameters:
        - $ref: "#/components/parameters/ObjectId"
      responses:
        "200":
          description: Dados do servico removidos com sucesso
        "404":
          description: Servico inexistente para o ID
  /requests:
    put:
      tags:
        - Pedidos
      summary: Cadastra os dados de um novo pedido na plataforma
      operationId: createRequest
      requestBody:
        description: Dado do pedido a ser cadastrado
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateRequestModel"
      responses:
        "200":
          description: Pedido criado com sucesso
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RequestModel"
        "400":
          description: Pedido requisitado possui dados invalidos
  /requests/{objectId}:
    get:
      tags:
        - Pedidos
      summary: Obtem os dados de um pedido existente por ID na plataforma
      parameters:
        - $ref: "#/components/parameters/ObjectId"
      responses:
        "200":
          description: Dados do pedido obtidos com sucesso
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RequestModel"
        "404":
          description: Pedido inexistente para o ID
      operationId: getRequestById
    patch:
      tags:
        - Pedidos
      summary: Atualiza os dados de um pedido existente por ID na plataforma
      parameters:
        - $ref: "#/components/parameters/ObjectId"
      responses:
        "200":
          description: Dados do pedido atualizados com sucesso
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RequestModel"
        "404":
          description: Pedido inexistente para o ID
      requestBody:
        description: Dado do pedido a ser atualizado
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateRequestModel"
      operationId: updateRequestById
  /user/{objectId}/requests:
    get:
      tags:
        - Utilidades
      summary: Obtem os dados de todos os pedidos de servicos feitos para um usuario prestador existente por ID
      parameters:
        - $ref: "#/components/parameters/ObjectId"
      responses:
        "200":
          description: Pedidos obtidos com sucesso para o ID do usuario prestador
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RequestListModel"
        "404":
          description: Usuario prestador invalido para o ID
      operationId: getRequestsByProviderId
  # /pets:
  #   put:
  #     tags:
  #       - Outros cadastros
  #     summary: Cadastra os dados de um novo pet na plataforma
  #     operationId: createPet
  #     requestBody:
  #       description: Dados do pet a ser cadastrado
  #       required: true
  #       content:
  #         application/json:
  #           schema:
  #             $ref: "#/components/schemas/PetData"
  #     responses:
  #       "200":
  #         description: Pet criado com sucesso
  #         content:
  #           application/json:
  #             schema:
  #               $ref: "#/components/schemas/Pet"
  #       "400":
  #         description: Pet requisitado possui dados invalidos
  # /pets/{objectId}:
  #   get:
  #     tags:
  #       - Outros cadastros
  #     summary: Obtem os dados de um pet existente por ID na plataforma
  #     operationId: getPetById
  #     parameters:
  #       - $ref: "#/components/parameters/ObjectId"
  #     responses:
  #       "200":
  #         description: Pet obtido com sucesso para o ID
  #         content:
  #           application/json:
  #             schema:
  #               $ref: "#/components/schemas/Pet"
  #       "404":
  #         description: Pet inexistente para o ID
  #   patch:
  #     tags:
  #       - Outros cadastros
  #     summary: Atualiza os dados de pet existente por ID na plataforma
  #     operationId: updatePetById
  #     parameters:
  #       - $ref: "#/components/parameters/ObjectId"
  #     requestBody:
  #       description: Dados do pet a ser atualizado
  #       required: true
  #       content:
  #         application/json:
  #           schema:
  #             $ref: "#/components/schemas/PetData"
  #     responses:
  #       "200":
  #         description: Pet obtido com sucesso para o ID
  #         content:
  #           application/json:
  #             schema:
  #               $ref: "#/components/schemas/Pet"
  #       "404":
  #         description: Pet inexistente para o ID
  #   delete:
  #     tags:
  #       - Outros cadastros
  #     summary: Remove os dados de um pet existente por ID na plataforma
  #     operationId: removePetById
  #     parameters:
  #       - $ref: "#/components/parameters/ObjectId"
  #     responses:
  #       "200":
  #         description: Pet removido com sucesso para o ID
  #       "404":
  #         description: Pet inexistente para o ID
components:
  schemas:
...
    ServiceModel:
      type: object
      properties:
        doc_id:
          type: string
          pattern: '^[a-fA-F\d]{24}$'
          example: 45cbc4a0e4123f6920000002
        name:
          type: string
          example: Tosa
        description:
          type: string
          example: Oferecemos servicos de tosa para cachorros de pequeno e medio porte.
        price:
          type: number
          format: double
          example: 89.90
        providerId:
          type: string
          pattern: '^[a-fA-F\d]{24}$'
          example: 45cbc4a0e4123f6920000002
        thumbnail:
          type: string
          format: byte
          example: U3dhZ2dlciByb2Nrcw==
      required:
        - "doc_id"
        - "name"
        - "price"
        - "providerId"
    RequestModel:
      type: object
      properties:
        doc_id:
          type: string
          pattern: '^[a-fA-F\d]{24}$'
          example: 45cbc4a0e4123f6920000002
        consumerId:
          type: string
          pattern: '^[a-fA-F\d]{24}$'
          example: 45cbc4a0e4123f6920000002
        serviceId:
          type: string
          pattern: '^[a-fA-F\d]{24}$'
          example: 45cbc4a0e4123f6920000002
        date:
          type: string
          format: datetime
          example: "1994-11-05T08:15:30-05:00"
        status:
          type: string
          example: Rejeitado
      required:
        - "doc_id"
        - "consumerId"
        - "serviceId"
        - "date"
    ServiceListModel:
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/ServiceModel"
      required:
        - "data"
    RequestListModel:
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/RequestModel"
      required:
        - "data"
...
    Pet:
      type: object
      properties:
        # doc_id:
        #   type: string
        #   pattern: '^[a-fA-F\d]{24}$'
        #   example: 45cbc4a0e4123f6920000002
        name:
          type: string
          example: "Scooby"
        species:
          type: string
          example: "Cachorro"
        race:
          type: string
          example: Dogue Alemao
        age:
          type: integer
          example: 5
        description:
          type: string
          example: Possui pedigree e foi comprado do criador "Chacara dos Dogues Alemaes".
        avatar:
          type: string
          format: byte
          example: U3dhZ2dlciByb2Nrcw==
      required:
        - "doc_id"
        - "name"
        - "species"
        - "race"
        - "age"
        - "description"
...
    UpdateServiceModel:
      type: object
      properties:
        name:
          type: string
          example: Tosa
        description:
          type: string
          example: Oferecemos servicos de tosa para cachorros de pequeno e medio porte.
        price:
          type: number
          format: double
          example: 89.90
        providerId:
          type: string
          pattern: '^[a-fA-F\d]{24}$'
          example: 45cbc4a0e4123f6920000002
      # required:
      #   - "name"
      #   - "price"
      #   - "providerId"
    UpdateRequestModel:
      type: object
      properties:
        consumerId:
          type: string
          pattern: '^[a-fA-F\d]{24}$'
          example: 45cbc4a0e4123f6920000002
        serviceId:
          type: string
          pattern: '^[a-fA-F\d]{24}$'
          example: 45cbc4a0e4123f6920000002
        date:
          type: string
          format: datetime
          example: "1994-11-05T08:15:30-05:00"
        thumbnail:
          type: string
          format: byte
          example: U3dhZ2dlciByb2Nrcw==
      # required:
      #   - "doc_id"
      #   - "consumerId"
      #   - "serviceId"
      #   - "date"
      # CreditCardData:
    #   type: object
    #   properties:
    #     code:
    #       type: string
    #       example: "5118598797832798"
    #     name:
    #       type: string
    #       example: "ARIEL S DOS SANTOS"
    #     expirationDate:
    #       type: string
    #       format: datetime
    #       example: "1994-11-05T08:15:30-05:00"
    #     cvv:
    #       type: string
    #       example: "183"
    #   required:
    #     - "code"
    #     - "name"
    #     - "expirationDate"
    #     - "cvv"
    # BankAccountData:
    #   properties:
    #     agency:
    #       type: string
    #       example: "60883"
    #     accountNumber:
    #       type: string
    #       example: "3413"
    #     digit:
    #       type: string
    #       example: "1"
    #   required:
    #     - "agency"
    #     - "accountNumber"
    #     - "digits"
    # AddressData:
    #   type: object
    #   properties:
    #     street:
    #       type: string
    #       example: Av. Paulista
    #     number:
    #       type: string
    #       example: "1001"
    #     complement:
    #       type: string
    #       example: apto 21
    #     city:
    #       type: string
    #       example: Sao Paulo
    #     state:
    #       type: string
    #       example: SP
    #     postcode:
    #       type: string
    #       example: "01311000"
    #   required:
    #     - "street"
    #     - "number"
    #     - "city"
    #     - "state"
    #     - "postcode"
    # PetData:
    #   type: object
    #   properties:
    #     name:
    #       type: string
    #       example: "Scooby"
    #     species:
    #       type: string
    #       example: "Cachorro"
    #     race:
    #       type: string
    #       example: Dogue Alemao
    #     age:
    #       type: integer
    #       example: 5
    #     description:
    #       type: string
    #       example: Possui pedigree e foi comprado do criador "Chacara dos Dogues Alemaes".
    #   required:
    #     - "name"
    #     - "species"
    #     - "age"
  parameters:
    ObjectId:
      name: objectId
      in: path
      description: ID do objeto
      required: true
      schema:
        type: string
        pattern: '^[a-fA-F\d]{24}$'
        example: 45cbc4a0e4123f6920000002